requirements.txt
python-pptx==1.0.2
boto3>=1.40.1
botocore>=1.40.1
python-docx==1.2.0
beautifulsoup4>=4.13.4
openpyxl>=3.1.2
fastmcp==3.2.0
pydantic>=2.11.5
PyYAML
pystache>=0.6.5
google-cloud-storage>=2.18.2
azure-storage-blob>=12.23.1
requests>=2.31.0
defusedxml>=0.7.1
python-dotenv>=1.2.1
